Split edit_spec into per-form files
